Separated WED analysis into separate program
Copied the WED analysis functions and other functions, parameters, and
variables which these require from the reconstruction program and
created a new program which just performs the WED analysis.

The path to the directory containing the subdirectories for the target
points and target volume files is defined automatically as the current
(i.e. program execution) directory or a user specified directory if this
is passed as a command line argument.  The paths to the target
points/volume and WED results are then defined relative to this main
directory using the predefined folder/file names specified in the header
file.  If the directory for the results does not exist, it is created.

The program is currently working, but the user defined specification of
the main directory has not yet been tested.  The program will be cleaned
up and tested thoroughly, after which the program will become available
to anyone that needs it.