Finished automated self contained WED program
Program is now complete and results have been verified against those
generated using the reconstruction program containing these same
functions as subroutines.

The IO directory containing separate subdirectories for the target
points, target volume, and WED results.  This is assumed to be the
current (execution) directory unless another directory is explicitly
passed to the program at execution through a command line argument.  The
names of these 3 subdirectories and the input/output files they contain
are predefined constants in the header file and their full paths are
defined relative to the main directory using these folder/file names.

The date of execution is also added to the WED results folder name so
and if it already exists, an integer value n is increased until adding
"_n" to the file name results in a unique directory name, preventing
results from the same and different days from accidentally being
overwritten and automatically organizing them.  Results can only be
discarded by manually deleting the folder they are in.  This is helpful
for debugging the program and comparing results from previous versions.